[PyTorch] Refactor Dispatcher to inline less code in fast path
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51163

The Dispatcher seems to have been in a precarious local
maximum: I tried to make several different changes to parameter
passing and ended up with regressions due to reduced inlining that
swamped any gains I might have gotten from the parameter passing
changes.

This diff reduces the amount of inline code on the fast path. It
should both reduce code size and provide a platform for making further
improvements to the dispatcher code.

It is a slight performance regression, but it unblocked the following
two diffs (which seem to get us back where we were) from landing.
ghstack-source-id: 120693163

Differential Revision: [D26091377](https://our.internmc.facebook.com/intern/diff/D26091377/)