mem-efficient learnable fake quantization
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49315

Update the learnable fake-quantization to use c++ and CUDA kernels, and resolve some issues on using it with pytorch DDP.
The updated quantization operator have a different gradient calculation for scale and zero_point when the output is at the endpoints of clamp operation. The updated quantization operator calculates the gradient according to the gradient of the `clamp` function. This behavior is consistent with the gradient calculation for non-learnable fake-quant ops.
ghstack-source-id: 120821868

Differential Revision: [D24479735](https://our.internmc.facebook.com/intern/diff/D24479735/)