[quant][refactor] Factor out MHA code dup from nn and nn.quantizable
Quantizable MHA shares some code with its FP equivalent.
This factors out common parts.

Test Plan:

Explicit testing is not available, as there is no functional tests for the nn.MHA.
However, the functional equivalence of this PR was tested across commits by comparing the numerics of the pre-change and post-change MHA. Similar was done for the quantizable counterpart.

ghstack-source-id: e2e150ee65a188d9d4d1b5a787bb18af231acbb6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51169